webshop/requirements.txt
asgiref==3.7.2
django==4.2.3
django-taggit==4.0.0
pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
